requirements.txt
asgiref==3.2.10
autopep8==1.5.4
boto3==1.17.58
botocore==1.20.58
CacheControl==0.12.6
cachetools==4.2.4
certifi==2020.12.5
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.4.2
Django==3.1
django-cors-headers==3.6.0
django-crispy-forms==1.10.0
django-embed-video==1.4.0
django-extensions==3.0.5
django-extra-fields==3.0.0
django-filter==2.3.0
django-import-export==2.3.0
django-jazzmin==2.2.7
django-phonenumber-field==5.2.0
django-rest-auth==0.9.5
django-silk==4.1.0
django-storages==1.11.1
django-urlconf-export==1.1.1
djangorestframework==3.11.1
djangorestframework-expander==0.2.3
djangorestframework-filters==1.0.0.dev2
djangorestframework-gis==0.15
djangorestframework-jwt==1.11.0
docutils==0.15.2
et-xmlfile==1.0.1
fcm-django==1.0.5
firebase-admin==5.0.3
google-api-core==2.1.0
google-api-python-client==2.24.0
google-auth==2.3.0
google-auth-httplib2==0.1.0
google-cloud-core==2.1.0
google-cloud-firestore==2.3.4
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
gprof2dot==2019.11.30
grpcio==1.41.0
gunicorn==20.1.0
httplib2==0.20.1
idna==2.10
inflect==5.3.0
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
MarkupPy==1.14
MarkupSafe==1.1.1
msgpack==1.0.2
odfpy==1.4.1
olefile==0.46
openpyxl==3.0.5
packaging==21.0
phonenumbers==8.12.31
Pillow==4.1.1
proto-plus==1.19.4
protobuf==3.18.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
Pygments==2.7.3
PyJWT==1.7.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.19.0
pytz==2020.4
PyYAML==5.3.1
requests==2.24.0
rsa==4.7.2
s3transfer==0.4.0
sentry-sdk==1.5.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
twilio==6.0.0
uritemplate==3.0.1
urllib3==1.25.11
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
